dom/push/test/test_permissions.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1038811: Push tests.

Any copyright is dedicated to the Public Domain.
http://creativecommons.org/licenses/publicdomain/

-->
<head>
  <title>Test for Bug 1038811</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8">
</head>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1038811">Mozilla Bug 1038811</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="text/javascript">

  function debug(str) {
  //  console.log(str + "\n");
  }

  function start() {
    return navigator.serviceWorker.register("worker.js" + "?" + (Math.random()), {scope: "."})
    .then((swr) => registration = swr);
  }

  function unregister() {
    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    }, function(e) {
      dump("Unregistering the SW failed with " + e + "\n");
    });
  }

  function setupPushNotification(swr) {
    var p = new Promise(function(res, rej) {
      swr.pushManager.subscribe().then(
        function(pushSubscription) {
          ok(false, "subscribe() should fail because no permission for push");
          res(swr);
        }, function(error) {
          ok(true, "subscribe() could not register for push notification");
          res(swr);
        }
        );
    });
    return p;
  }

  function getEndpoint(swr) {
    var p = new Promise(function(res, rej) {
      swr.pushManager.getSubscription().then(
        function(pushSubscription) {
          ok(false, "getSubscription() should fail because no permission for push");
          res(swr);
        }, function(error) {
          ok(true, "getSubscription() could not register for push notification");
          res(swr);
        }
        );
    });
    return p;
  }

  function runTest() {
    start()
    .then(setupPushNotification)
    .then(getEndpoint)
    .then(unregister)
    .catch(function(e) {
      ok(false, "Some test failed with error " + e);
    }).then(SimpleTest.finish);
  }

  SpecialPowers.addPermission('push', false, document);
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.push.enabled", true],
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
    ]}, runTest);
  SimpleTest.waitForExplicitFinish();

</script>
</body>
</html>
